pyproject.toml
[project]
name = "spotifymanager"
version = "0.1.0"
description = "AI-powered Spotify library manager with dynamic playlists"
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.24.0",
    "pylast>=5.3.0",
    "sqlalchemy>=2.0",
    "thefuzz[speedup]>=0.22.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]
sprint2 = [
    "numpy>=1.26",
    "faiss-cpu>=1.8",
    "scikit-learn>=1.4",
]
sprint4 = [
    "musicbrainzngs>=0.7",
]
sprint5 = [
    "hdbscan>=0.8",
    "umap-learn>=0.5",
    "lightgbm>=4.3",
    "anthropic>=0.30",
]
sprint6 = [
    "discord.py>=2.4",
]

[project.scripts]
spotifymanager = "src.interface.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
